==> 01_FMA_data_preprocessing.ipynb <==
# Preprocessing all of the raw files from the Free Music Archive repo.

%matplotlib inline

import os

import IPython.display as ipd
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import sklearn as skl
import sklearn.utils, sklearn.preprocessing, sklearn.decomposition, sklearn.svm
import librosa
import librosa.display

# Reading in the raw files from the Free Music Archive repo so that I can clean them up a bit.

tracks = pd.read_csv('tracks.csv')
genres = pd.read_csv('genres.csv')
features = pd.read_csv('features.csv')

tracks.head()

# You can see that there's some weirdness where the first two rows contain additional column information. 
# We need to clean that up.

for_review = tracks.head()
for_review

for_review.to_csv("for_review.csv")

tracks.drop(['album', 'album.1', 'album.3', 'album.4', 'album.6', 'album.7', 'album.8', 'album.9', \
            'artist.2', 'artist.3', 'artist.4', 'artist.5', 'artist.6', 'artist.8', 'artist.9', 'artist.10', \
            'artist.11', 'artist.13', 'artist.14', 'artist.15', 'artist.16', 'track', 'track.1', 'track.2', \
            'track.3', 'track.6', 'track.8', 'track.9', 'track.10', 'track.11', 'track.12', 'track.13', 'track.15', \
            'track.17', 'track.18'], axis=1, inplace=True)

tracks.columns = ['track_id', 'album_date_released', 'album_id', 'album_title', 'album_tracks', 'album_type', \
                  'artist_year_begin', 'artist_year_end', 'artist_id', 'artist_name', 'set_split', 'set_subset', \
                  'track_date_recorded', 'track_duration', 'track_genre_top', 'track_listens', 'track_number', \
                  'track_title']

tracks.head()

# looks much better!

# Now we need to drop the first two rows that don't have actual data:

tracks.drop([0,1], inplace=True)
tracks.head()

tracks.to_csv("tracks_updated.csv")

features.head()

# These are a lot better, thankfully, although I'll have to drop the first three rows and rename the first column.
# The first three rows do have more information about the variables, but they're not data and I don't think
# it's worth renaming all the columns.

for_review = features.head()
for_review.to_csv("for_review.csv")

features.rename(index=str, columns={"feature": "track_id"}, inplace = True)
features.drop(features.index[[0,1,2]], inplace = True)
features.head()

features.to_csv("features_updated.csv")

echonest = pd.read_csv('echonest.csv')
echonest.head()

# Same deal as above. A few of these columns are not useful and will be dropped; others could be renamed
# but it's not important so I won't. Many of them shouldn't need to be renamed anyway.

# Update -- I didn't end up using these because this is a relatively small subset of the tracks (~10%)

for_review = echonest.head()
for_review.to_csv("for_review.csv")

echonest.rename(index=str, columns={"Unnamed: 0": "track_id"}, inplace = True)
echonest.drop(['echonest.8', 'echonest.9', 'echonest.10', 'echonest.11', 'echonest.12', 'echonest.13', 'echonest.14',
              'echonest.15', 'echonest.16', 'echonest.17', 'echonest.18', 'echonest.19', 'echonest.20', \
              'echonest.21', 'echonest.22', 'echonest.23', 'echonest.24'], axis=1, inplace=True)
echonest.drop(['0','1','2'], inplace = True)
echonest.head()

echonest.to_csv("echonest_updated.csv")
