==> 02_FMA_data_cleaning.ipynb <==
import pandas as pd
from sqlalchemy import create_engine

# Pulling the tracks dataset out of postgresql on an AWS EC2 instance -- I cleaned this a bit before loading into
# SQL but it still needs a little cleaning and preparation. Note that I removed the password and IP address
# information from the line below.

cnx = create_engine('postgresql://username:password@IPAddress:5432')

# Testing pulling data from AWS:

pd.read_sql_query('''SELECT * FROM tracks LIMIT 5''', cnx)

# It works!

# Pulling over the rest of our tracks data from AWS

tracks = pd.read_sql_query('''SELECT * FROM tracks''', cnx)

# Now we'll read in the rest of our data (far too many columns to load into SQL)

features = pd.read_csv('features_updated.csv')

tracks.head()

tracks.drop(['unnamed_0'], axis=1, inplace=True)
tracks.head()

tracks.track_id.count()

# Since the "top" genre is our dependent variable, let's see what we're dealing with.

tracks.track_genre_top.value_counts()

# Looks like a bunch are missing -- let's check all of the variables.

tracks.isna().sum()

# Looks like the artists' beginning and end years and the dates tracks were recorded are mostly missing so we will
# drop them. Also we will have to drop the 56,976 albums missing a top genre -- we might use them in a future
# project but this is going to be complicated enough already!

tracks.drop(['artist_year_begin', 'artist_year_end', 'track_date_recorded'], axis=1, inplace=True)
tracks = tracks[tracks.track_genre_top.notna()]
print(tracks.track_id.count())
tracks.isna().sum()

# Next let's check out the features data we read in from CSV and then merge the two.

features.head()

features.drop(['Unnamed: 0'], axis=1, inplace=True)
features.head()

features.isna().sum()

# sweet!

# Now let's merge features into the tracks dataframe.

merged_df = tracks.merge(features, on = 'track_id')
merged_df.shape
# Merge looks good!

# Rather than create my own train/test split I'm going to use the split's they've defined, because if I choose to 
# limit a model to the small or medium data sets the train/test splits are already defined within each of those
# subsets as well as the whole file. 

# Additionally, I'm going to use their validation split instead of running KFold or another cross-validation method
# for two reasons. First, I'll be upsampling within the training set and would prefer that the validation set
# not be upsampled as well. Second, the train/validation/test splits in the data are already stratified and also
# make sure that all songs from a particular artist appear in only one of the three sets. Having songs from the
# same artist appear in both test and train may reward overfitting since they are likely to be very similar.

# The small subset includes 8,000 tracks evenly divided among 8 genres, and the medium subset includes an additional
# 17,000 tracks to produce an unbalanced set of tracks that all have a single top genre, excellent metadata, and
# are relatively popular. The remaining tracks in the large dataset still have ample metadata for my purposes
# as well as a single top genre, so I don't have any a priori reason to exclude them.

print(merged_df.set_split.value_counts())
print(merged_df.set_subset.value_counts())
print(pd.crosstab(merged_df['track_genre_top'], merged_df['set_subset']))

# Now to turn the genres into numeric categories for modeling purposes:

merged_df['track_genre_top'] = merged_df['track_genre_top'].map({'Rock' : 0, 'Experimental' : 1, \
    'Electronic' : 2, 'Hip-Hop' : 3, 'Folk' : 4, 'Pop' : 5, 'Instrumental' : 6, 'International' : 7, \
    'Classical' : 8, 'Jazz' : 9, 'Old-Time / Historic' : 10, 'Spoken' : 11, 'Country' : 12, \
    'Soul-RnB' : 13, 'Blues' : 14, 'Easy Listening' : 15})

merged_df.track_genre_top.value_counts()

# It's "don't be a hero" time and so I'm dropping the 7 genres that have less than 1,000 tracks in my dataset.
# This still leaves me with 9 genres, which is plenty complicated enough.

merged_df.drop(merged_df[merged_df.track_genre_top > 8].index, inplace = True)
merged_df.track_genre_top.value_counts()

# Let's break out into X and y train, validate, and test:

X_train = merged_df[merged_df.set_split == "training"]
X_val = merged_df[merged_df.set_split == "validation"]
X_test = merged_df[merged_df.set_split == "test"]

y_train = X_train['track_genre_top']
y_val = X_val['track_genre_top']
y_test = X_test['track_genre_top']


X_train.drop(['track_genre_top', 'album_id', 'album_title', 'artist_id', 'artist_name', 'set_split', \
              'set_subset', 'track_title', 'album_date_released', 'album_type'], axis=1, inplace=True)
X_val.drop(['track_genre_top', 'album_id', 'album_title', 'artist_id', 'artist_name', 'set_split', \
              'set_subset', 'track_title', 'album_date_released', 'album_type'], axis=1, inplace=True)
X_test.drop(['track_genre_top', 'album_id', 'album_title', 'artist_id', 'artist_name', 'set_split', \
              'set_subset', 'track_title', 'album_date_released', 'album_type'], axis=1, inplace=True)

print(X_train.shape)
print(X_val.shape)
print(X_test.shape)
print(y_train.shape)
print(y_val.shape)
print(y_test.shape)

# Let's reduce some dimensions! We'll apply PCA to our training set and then apply the same transformations
# to X_val and X_test. Let's try reducing our 500-ish dimensions down to 15:

from sklearn.decomposition import PCA

pca = PCA(n_components=15)
pca.fit(X_train)

pca.explained_variance_ratio_


#Looks reasonable -- the explained variance ratios get small pretty quickly.

# Now we transform all of our X sets using the PCA that is based only on X_train (so there is no feature bleed):

X_train_tr = pca.transform(X_train)
X_val_tr = pca.transform(X_val)
X_test_tr = pca.transform(X_test)

print(X_train_tr.shape)
print(X_val_tr.shape)
print(X_test_tr.shape)

# Now let's oversample within our training dataset using SMOTE because I like the idea of it making up
# fake songs from smaller genres rather than just copying existing songs. We are using the PCA-transformed
# version of X_train.

from imblearn.over_sampling import SMOTE

sm = SMOTE(random_state=2)
X_train_res, y_train_res = sm.fit_sample(X_train_tr, y_train.ravel())

print(X_train_res.shape)
print(y_train_res.shape)

# Now save both the oversampled X_train and y_train as well as the non-oversampled versions, plus our validation
# and test sets:

pd.DataFrame(X_train_tr).to_csv("X_train.csv")
pd.DataFrame(X_train_res).to_csv("X_train_oversmp.csv")
pd.DataFrame(X_val_tr).to_csv("X_val.csv")
pd.DataFrame(X_test_tr).to_csv("X_test.csv")

y_train.to_csv("y_train.csv")
pd.DataFrame(y_train_res).to_csv("y_train_oversmp.csv")
y_val.to_csv("y_val.csv")
y_test.to_csv("y_test.csv")

# Making a top-3 genre only version so I can also model only on those genres:

top3_df = merged_df.drop(merged_df[merged_df.track_genre_top > 2].index)
print(top3_df.track_genre_top.value_counts())

X_train_top3 = top3_df[top3_df.set_split == "training"]
X_val_top3 = top3_df[top3_df.set_split == "validation"]
X_test_top3 = top3_df[top3_df.set_split == "test"]

y_train_top3 = X_train_top3['track_genre_top']
y_val_top3 = X_val_top3['track_genre_top']
y_test_top3 = X_test_top3['track_genre_top']


X_train_top3.drop(['track_genre_top', 'album_id', 'album_title', 'artist_id', 'artist_name', 'set_split', \
              'set_subset', 'track_title', 'album_date_released', 'album_type'], axis=1, inplace=True)
X_val_top3.drop(['track_genre_top', 'album_id', 'album_title', 'artist_id', 'artist_name', 'set_split', \
              'set_subset', 'track_title', 'album_date_released', 'album_type'], axis=1, inplace=True)
X_test_top3.drop(['track_genre_top', 'album_id', 'album_title', 'artist_id', 'artist_name', 'set_split', \
              'set_subset', 'track_title', 'album_date_released', 'album_type'], axis=1, inplace=True)

print(X_train_top3.shape)
print(X_val_top3.shape)
print(X_test_top3.shape)
print(y_train_top3.shape)
print(y_val_top3.shape)
print(y_test_top3.shape)

pca_top3 = PCA(n_components=15)
pca_top3.fit(X_train_top3)

X_train_top3 = pca.transform(X_train_top3)
X_val_top3 = pca.transform(X_val_top3)
X_test_top3 = pca.transform(X_test_top3)

print(X_train_top3.shape)
print(X_val_top3.shape)
print(X_test_top3.shape)

pd.DataFrame(X_train_top3).to_csv("X_train_top3.csv")
pd.DataFrame(X_val_top3).to_csv("X_val_top3.csv")
pd.DataFrame(X_test_top3).to_csv("X_test_top3.csv")

y_train_top3.to_csv("y_train_top3.csv")
y_val_top3.to_csv("y_val_top3.csv")
y_test_top3.to_csv("y_test_top3.csv")

X_train.head()

# For use in our prediction app -- so we can fill these in with median values in uploaded tracks.

print(X_train.track_id.median())
print(X_train.album_tracks.median())
print(X_train.track_duration.median())
print(X_train.track_listens.median())
print(X_train.track_number.median())

# We also need to pickle our PCA transformation for use in our prediction app.

import pickle

pickle.dump(pca, open('PCA_transform.pkl', 'wb'))
